feat: add support for layered architecture

+++ src/package-utils/setup-helpers.ts
@@ -2,6 +2,7 @@ import { execa } from "execa";
 import {
   installingPrismaClientWithPM,
   installingPrismaCLIWithPM,
+  type PackageManager,
 } from "./detect-pm";
 import {
   log,
@@ -39,9 +40,12 @@ export async function isPrismaCLIInstalled(
   }
 }
 
-export async function installPrismaCLI(directory: string) {
+export async function installPrismaCLI(
+  directory: string,
+  packageManager?: PackageManager,
+) {
   try {
-    const installCmd = installingPrismaCLIWithPM();
+    const installCmd = installingPrismaCLIWithPM(directory, packageManager);
 
     await execa(installCmd.pm, installCmd.command, {
       cwd: directory,
@@ -72,19 +76,19 @@ export async function initPrisma({
   provider = "sqlite",
   datasourceUrl,
 }: PrismaInitOptions) {
-  const command = ["prisma", "init", "--datasource-provider"];
+  const commandArgs = ["prisma", "init", "--datasource-provider"];
 
-  command.push(provider);
+  commandArgs.push(provider);
 
   if (datasourceUrl) {
-    command.push("--url");
-    command.push(datasourceUrl);
+    commandArgs.push("--url");
+    commandArgs.push(datasourceUrl);
   }
 
   try {
     log(PREDEFINED_LOG_MESSAGES.initPrisma.action);
 
-    const { stdout: initializePrisma } = await execa("npx", command, {
+    const { stdout: initializePrisma } = await execa("npx", commandArgs, {
       cwd: directory,
     });
 
@@ -120,23 +124,26 @@ export async function writeToSchema(prismaSchemaPath: string) {
       return false;
     }
 
-    const addModel = `
-            model User {
-              id    Int     @id @default(autoincrement())
-              email String  @unique
-              name  String?
-              posts Post[]
-            }
+    const addModel = `\
+model User {
+  id    Int     @id @default(autoincrement())
+  email String  @unique
+  name  String?
+  posts Post[]
+}
 
-            model Post {
-              id        Int     @id @default(autoincrement())
-              title     String
-              content   String?
-              published Boolean @default(false)
-              author    User    @relation(fields: [authorId], references: [id])
-              authorId  Int
-            }
-          `;
+model Post {
+  id        Int     @id @default(autoincrement())
+  title     String
+  content   String?
+  published Boolean @default(false)
+  author    User    @relation(fields: [authorId], references: [id])
+  authorId  Int
+}
+`;
+
+    // Don't bother adding the models if they already exist.
+    if (existingSchema.trim().includes(addModel.trim())) return;
 
     const updatedSchema = `${existingSchema.trim()}\n\n${addModel}`;
     writeFileSync(prismaSchemaPath, updatedSchema);
@@ -145,13 +152,17 @@ export async function writeToSchema(prismaSchemaPath: string) {
   }
 }
 
-export async function runMigration(directory: string) {
+export async function runMigration(directory: string, schemaPath: string[]) {
   try {
     log(PREDEFINED_LOG_MESSAGES.runMigration.action);
 
-    await execa("npx", ["prisma", "migrate", "dev", "--name", "init"], {
-      cwd: directory,
-    });
+    await execa(
+      "npx",
+      ["prisma", "migrate", "dev", "--name", "init"].concat(schemaPath),
+      {
+        cwd: directory,
+      },
+    );
     logSuccess(PREDEFINED_LOG_MESSAGES.runMigration.success);
     return true;
   } catch (err) {
@@ -162,24 +173,30 @@ export async function runMigration(directory: string) {
   }
 }
 
-export async function formatSchema(directory: string) {
+export async function formatSchema(directory: string, schemaPath: string[]) {
   try {
     log(PREDEFINED_LOG_MESSAGES.formatSchema.action);
-    await execa("npx", ["prisma", "format"], { cwd: directory });
+    await execa("npx", ["prisma", "format"].concat(schemaPath), {
+      cwd: directory,
+    });
   } catch {
     logError(PREDEFINED_LOG_MESSAGES.formatSchema.error);
   }
 }
 
-export async function generateClient(
+export async function installPrismaClient(
   directory: string,
   installPrismaClient: boolean = true,
+  packageManager?: PackageManager,
 ) {
   log(PREDEFINED_LOG_MESSAGES.generatePrismaClient.action);
 
   if (installPrismaClient) {
     try {
-      const installCmd = installingPrismaClientWithPM();
+      const installCmd = installingPrismaClientWithPM(
+        directory,
+        packageManager,
+      );
 
       await execa(installCmd.pm, installCmd.command, {
         cwd: directory,
@@ -192,32 +209,44 @@ export async function generateClient(
       // log(error);
     }
   }
+}
 
+export async function generatePrismaClient(
+  directory: string,
+  prismaSchemaPath: string[],
+  verboseLog: boolean = false,
+) {
   try {
     const { stdout: generateClient } = await execa(
       "npx",
-      ["prisma", "generate"],
+      ["prisma", "generate"].concat(prismaSchemaPath),
       { cwd: directory },
     );
 
     log("\n" + generateClient.split("\n").slice(0, 4).join("\n") + "\n");
-
-    // log(generateClient);
   } catch (err) {
     logError(PREDEFINED_LOG_MESSAGES.generatePrismaClient.error);
-    // log(err);
+    if (verboseLog) {
+      log(err);
+    }
   }
 }
 
-export async function installStudio(directory: string) {
+export async function installStudio(directory: string, schemaPath?: string) {
   try {
-    const { spawn } = require("child_process");
-
     log(PREDEFINED_LOG_MESSAGES.installStudio.action);
 
-    await spawn("npx", ["prisma", "studio", "--browser", "none"], {
-      cwd: directory,
-    });
+    const schemaLocation = schemaPath ? ["--schema", schemaPath] : [];
+
+    const subprocess = execa(
+      "npx",
+      ["prisma", "studio", "--browser", "none"].concat(schemaLocation),
+      {
+        cwd: directory,
+      },
+    );
+
+    subprocess.unref();
 
     logSuccess(PREDEFINED_LOG_MESSAGES.installStudio.success);
 
@@ -230,11 +259,12 @@ export async function installStudio(directory: string) {
 }
 
 export async function writeClientInLib(path: string) {
-  const existingContent = existsSync(path);
+  const existingContent = existsSync(`${path}/lib/prisma.ts`);
 
   try {
     if (!existingContent) {
-      const prismaClient = `import { PrismaClient } from '@prisma/client'
+      const prismaClient = `\
+import { PrismaClient } from '@prisma/client'
 
 const prismaClientSingleton = () => {
   return new PrismaClient()
@@ -251,16 +281,16 @@ export default prisma
 if (process.env.NODE_ENV !== 'production') globalThis.prismaGlobal = prisma
 `;
 
-      if (!existsSync("lib")) {
-        mkdirSync("lib");
+      if (!existsSync(`${path}/lib`)) {
+        mkdirSync(`${path}/lib`);
       }
 
-      if (existsSync("lib/prisma.ts")) {
+      if (existsSync(`${path}/lib/prisma.ts`)) {
         log(PREDEFINED_LOG_MESSAGES.writeClientInLib.found);
         return;
       }
 
-      writeFileSync("lib/prisma.ts", prismaClient);
+      writeFileSync(`${path}/lib/prisma.ts`, prismaClient);
 
       logSuccess(PREDEFINED_LOG_MESSAGES.writeClientInLib.success);
     }
